Move SearchSource interface to Search file.
